Use the term "regular" instead of "norep"
The term "regular" means that a scheduled item
should be shown in the agenda only the first/next time that
it occurs.

For example, if we have a regular mtg @ 10:00 each day,
I don't want to see 5 "Meeting" entries in my agenda every week.

I just want to be reminded of the *next* occurrence of this regularly
occurring thing.